[P2P-199] fix: 코스(장소) 상세 - 댓글 간 갯수 상이한 점 수정

+++ src/main/java/com/prgrms/p2p/domain/course/service/CourseQueryService.java
@@ -33,7 +33,7 @@ public class CourseQueryService {
         .orElseThrow(() -> new NotFoundException("존재하지 않는 코스입니다."));
     Boolean isLiked = likeRepository.existsByUserIdAndCourse(userId, course);
     Boolean isBookmarked = bookmarkRepository.existsByUserIdAndCourse(userId, course);
-    Integer comments = courseCommentRepository.countByCourse(course);
+    Long comments = courseCommentRepository.countByCourse(course);
     return CourseConverter.ofDetail(course, isLiked, isBookmarked, comments);
   }
